pyproject.toml
[project]
name = "pacegenie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.1.3",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.18",
    "openevals>=0.1.3",
    "pandas>=2.0",
    "pgvector>=0.4.2",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.40",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.18",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
